wip: clean comments and initialize app build
