+++ lib/app.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class App extends StatelessWidget {
+  const App({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Good App',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: Scaffold(
+        appBar: AppBar(
+          title: const Text('App'),
+          centerTitle: true,
+        ),
+        body: Container(),
+      ),
+    );
+  }
+}
